+++ src/app/_components/DisplayProducts.tsx
@@ -1,8 +1,8 @@
 import { Card, CardContent, Grid } from '@mui/material';
 import { numbers, cardInfo } from '@/src/data/dummy';
 import Image from 'next/image';
-import ProductColors from './ProductColors';
-import ProductInfo from './ProductInfo';
+import ProductColors from '../../components/ui/ProductColors';
+import ProductInfo from '../../components/ui/ProductInfo';
 
 function DisplayProducts() {
   return (
