+++ src/components/ui/ProductInfo.tsx
